two sensors working :)

+++ benchmarks/source/superh/port/devsensor.h
@@ -1,2 +1,2 @@
-float	devsignal_read(uint32_t which);
+double	devsignal_read(int which);
 void	devsignal_write(double value);
